+++ Encoding/x264Enc.vb
@@ -108,7 +108,7 @@ Public Class x264Enc
         End If
 
         script.Filters.Add(New VideoFilter("aaa", "aaa", code))
-        script.Path = p.TempDir + p.TargetFile.Base + "_CompCheck." + script.FileType
+        script.Path = (p.TempDir + p.TargetFile.Base + "_CompCheck." + script.FileType).ToShortFilePath
         script.Synchronize()
 
         Log.WriteLine(BR + script.GetFullScript + BR)
@@ -1114,7 +1114,7 @@ Public Class x264Params
 
                 args += " --output NUL " + input
             Else
-                args += " --output " + targetPath.Escape + " " + input
+                args += " --output " + targetPath.ToShortFilePath.Escape + " " + input
             End If
         End If