+++ General/LogBuilder.vb
@@ -151,9 +151,9 @@ Public Class LogBuilder
         If proj.SourceFile = "" Then
             Return Folder.Temp + "staxrip.log"
         ElseIf proj.TempDir = "" Then
-            Return g.GetPath(proj.SourceFile.Dir, proj.SourceFile.Base, "_staxrip", "log")
+            Return proj.SourceFile.Dir + proj.SourceFile.Base + "_staxrip.log"
         Else
-            Return g.GetPath(proj.TempDir, proj.TargetFile.Base, "_staxrip", "log")
+            Return proj.TempDir + proj.TargetFile.Base + "_staxrip.log"
         End If
     End Function
 End Class